requirements.txt
azure-cognitiveservices-vision-computervision==0.9.0
pandas==1.4.3
streamlit==1.10.0